Stop allowing non-terminal padding bytes within byte sequences
Previously, due to the way the data-decoding crate works, `=` was
ignored anywhere it occurred, rather than only being treated as a
padding byte. Unfortunately, this wasn't detected before due to the lack
of specification-test coverage prior to
https://github.com/httpwg/structured-field-tests/pull/94.

In this change we replace data-decoding with the base64 crate, allowing
us to fix this behavior, which should require a version bump.