Discard insignificant digits from decimals during serialization
Per Step 9 of https://httpwg.org/specs/rfc8941.html#ser-decimal, only
the significant digits of a decimal's fractional component are
serialized.

Previously, this caused a decimal like `123.10` to be serialized
verbatim rather than `123.1`. This was not caught by the specification
tests because the parsing function happens to normalize the parsed value
itself, and unfortunately the serialization-only specification tests
cannot cover this case because there is no guarantee that a JSON number
`123.10` will retain the trailing 0 on parsing.

In any case, this crate should not assume that a given decimal is
already normalized.