Simplify ref serializers
Arguably the explicit nesting here is easier to understand than the
builder-chaining of the previous setup, especially now that the method
names are shortened, and it's unclear how frequently user code would
take advantage of the builder-chaining in the first place.

This also makes dynamic misuse of the parameter-vs-value ordering
statically impossible.