Rework the lifetimes on visitors
This changes the trait definitions to break the 1:1 correlation between
the lifetime of the input and the lifetime of the various visitor
traits.  Instead, the lifetime of any visitor that is returned is tied
back to the visitor that created it.

This enables the sorts of implementation pattern that you see in the
tests here.  It had no effect on the internal use of visitors, other
than to change the method declarations.

Closes #179.