Change visitor signatures to return Option
This doesn't change them all.  Only those where ignoring makes sense
have this.  This covers:

* Parameters can be ignored entirely.
* Dictionary entries can be ignored after seeing the name.

Other cases are where values are expected, so these are not trivially
ignorable.  For instance, once committed to processing an item, that
item can only be ignored by returning `Ignored`; most of these cases are
ones where the parsing code is committed to proceeding.

The true value of this is where you have a parser that is conditionally
processing dictionary keys.  That parser doesn't need to concern itself
with how the visitor it returns handles stuff it isn't interested in.

Closes #183.