Fix clippy::pedantic errors (#186)
These were not uniformly easy.  There are a couple of cases where the
public API changes slightly, so this should be a new dot release.

Those changes are mostly through the inclusion of `#[must_use]` in a bunch of places.
Using code will then get warnings when they throw values away.

The main problem I found there was that the serializer code relied on a
Drop implementation in a couple of places.  I personally don't like that
very much, so I added `#[must_use]` anyway.  Using code probably isn't
exposed to this detail, even if they using custom serialization
arrangements, as it was only test code that had to add a bunch of
useless bindings to deal with this.  Still, this is worth noting.